Add mednafen support
Because VBA-M has decided to be speed racer in Docker, install mednafen
in the Docker image, and configure it to use VBA-M keybinds (so we have
consistent keybinds across emulators).

This is our fallback in case we can't figure other things out, at least.
Good to have in our back pocket.